Split playlist and traffic log stores
- Move traffic log data into its own store so it's only loaded by the /playlist view
- Rename playlistStore for clarity with two stores in the same "module" now
- Check if store is loading before getting missing entries, just in case another tab is also trying to load them
- Fix issues with adding an entry when there are two spots in the same slot and with re-opening the group after it's read
