+++ client/src/components/music/TrackList.vue
@@ -35,7 +35,10 @@
                   <ArtistLink :artist="track.track_artist" />
                 </span>
               </div>
-              <div v-if="track.pronunciation" class="text-muted fw-light fst-italic">
+              <div
+                v-if="track.pronunciation"
+                class="text-muted fw-light fst-italic"
+              >
                 {{ track.pronunciation }}
               </div>
               <div class="mb-1">
